Fix errors in packages tests
[Linux/macOS] Fix data races:
Added channel to watcher structure so that
the watcher.Close function can safely wait
until the goroutine is done and avoid data
races that could occur with just the mutex locks.

[macOS] Fix json_test and watcher_test in packages:
The way FSEvents handles events we need to use the
flushDir function or the notify framework won't
generate events for files if the directory had already
files watched inside.
The flushDir was being called at the end of the Watch
method and therefore it wasn't executing in some cases
where the method was returning before the call.