Changed back to using fsEvents
It turns out that the fsevent backend for the
notify library wasn't the cause of the tests
failing.

These proposed changes fix the errors in the
packages module.

On macOs there's a random panic in the backend
tests when using the oniguruma library but
I couldn't track it down yet and I don't want
to delay this fix anymore. The panic also
occurs in the master branch so it's not caused
by these changes.